requirements.txt
python-dotenv
pandas
numpy
google-generativeai
fastapi 
fastapi[standard]
uvicorn
